add timeout between boradcasting files
it happens that when broadcasting tons of files
the messages on the receiving side could not be deserialized
correctly. i guess this is something upd specific. for now this is a horrible fix.
it would be an option only broadcast ip addr and port and then other
instances could request file lists over tcp to prevent loss of data