allow mfsync get to request by directory
when using 'mfsync get --request' mfsync does not only retreive files
with the exact same filename but all files that start with the same
name.
this way it is possible to request all files within a folder by just
specifying the foldername because the
filenames (paths) will start with the folder name.