create ~/.mfsync config directory and add --config flag
if no --key-file flag is specified mfsync will create ~/.mfsync and
store the key.bin within that directory.
mfsync will also check for a config.json file within that directory and
use it if it exists.
as alternative users can specify a path to a config file using --config

the config file currently can be used to list allowed public keys.
example:
{
  "trustedKeys": [
    "8FA14E51FCD446F1E0B3AEA53751CCF26FF5ADD271FAA3AA4318120B23CAA120",
    "7D477C5FEBC40A655962A262EAE7F7D5FF169F6A767C945F99CD0E9070C50002",
    "CC6B53A585170C2CC29B4BDF064FDEFD79A1CE2DB96186176C9DB7D900CC2B00"
  ]
}